refactor: change to commomjs module system
