+++ package.json
@@ -5,7 +5,6 @@
   "main": "dist/index.js",
   "module": "dist/index.js",
   "types": "dist/index.d.ts",
-  "type": "module",
   "files": [
     "dist",
     "README.md",
@@ -13,7 +12,7 @@
   ],
   "scripts": {
     "build": "rollup -c",
-    "dev": "node --env-file=.env --watch -r ts-node/register src/index.ts",
+    "dev": "node --env-file=.env --watch -r ts-node/register src/server.ts",
     "start": "node --env-file=.env dist/index.js",
     "test": "vitest",
     "test:coverage": "vitest --coverage"
