+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './container'
 export * from './decorators'
 export * from './models'
+export * from './utils'
