+++ src/utils/merge-schema.test.ts
@@ -1,4 +1,4 @@
-import { mergeSchema } from '@/utils/merge-schema'
+import { mergeSchema } from '@/utils'
 
 describe('mergeSchema', () => {
   test('should merge schemas correctly', () => {
