+++ tsconfig.json
@@ -1,8 +1,7 @@
 {
   "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "moduleResolution": "node",
+    "target": "ES2020",
+    "module": "CommonJS",
     "esModuleInterop": true,
     "removeComments": true,
     "emitDecoratorMetadata": true,
